+++ packages/cmd/src/CommandPreparer.ts
@@ -450,6 +450,14 @@ export class CommandPreparer {
             type: "boolean"
         });
 
+        definition.options.push({
+            name: Constants.CURRENT_WORKING_DIRECTORY_OPTION,
+            aliases: [Constants.CURRENT_WORKING_DIRECTORY_OPTION_ALIAS],
+            group: Constants.GLOBAL_GROUP,
+            description: "Alternate current working directory",
+            type: "string"
+        });
+
         /**
          * Append any profile related options
          */