+++ packages/cmd/src/yargs/CommandYargs.ts
@@ -10,7 +10,7 @@
 */
 
 import { Arguments, Argv, Options } from "yargs";
-import { isNullOrUndefined } from "util";
+import { isNullOrUndefined, inspect } from "util";
 import { Constants } from "../../../constants";
 import { IYargsResponse } from "./doc/IYargsResponse";
 import { AbstractCommandYargs, YargsCommandCompleted } from "./AbstractCommandYargs";
@@ -19,6 +19,7 @@ import { ICommandDefinition } from "../doc/ICommandDefinition";
 import { CommandProcessor } from "../CommandProcessor";
 import { ICommandResponse } from "../../src/doc/response/response/ICommandResponse";
 import { CommandResponse } from "../../src/response/CommandResponse";
+import { ImperativeConfig } from "../../../utilities";
 
 /**
  * Define an Imperative Command to Yargs. A command implies that an implementation is present (differs from a "group")
@@ -146,7 +147,7 @@ export class CommandYargs extends AbstractCommandYargs {
                                     `Error in command ${this.definition.name}`,
                                     "command handler invoked", errorResponses);
                                 this.log.error(`Error in command ${this.definition.name}`);
-                                this.log.error(require("util").inspect(errorResponses));
+                                this.log.error(inspect(errorResponses));
                                 commandExecuted(argsForHandler, response);
                             });
                     } else {
@@ -224,7 +225,7 @@ export class CommandYargs extends AbstractCommandYargs {
                             }),
                             profileManagerFactory: this.profileManagerFactory,
                             rootCommandName: this.rootCommandName,
-                            commandLine: this.commandLine,
+                            commandLine: ImperativeConfig.instance.commandLine,
                             envVariablePrefix: this.envVariablePrefix,
                             promptPhrase: this.promptPhrase
                         }).invoke({
@@ -256,7 +257,8 @@ export class CommandYargs extends AbstractCommandYargs {
                     } catch (processorError) {
                         const response = new CommandResponse({
                             silent: false,
-                            responseFormat: (printJson) ? "json" : "default"
+                            responseFormat: (printJson) ? "json" : "default",
+                            stream: argsForHandler.stream
                         });
                         response.failed();
                         response.console.errorHeader("Internal Command Error");