+++ packages/utilities/src/TextUtils.ts
@@ -219,7 +219,6 @@ export class TextUtils {
      */
     public static wordWrap(text: string, width?: number,
                            indent: string = "", hardWrap: boolean = false): string {
-        const yargs = require("yargs");
         const wrappedText = require("wrap-ansi")(text, this.getRecommendedWidth(width), {hard: hardWrap});
         return wrappedText.split(/\n/g).map((line: string) => {
             if (line.length === 0) {